+++ deploy/scripts/src/spot/request/types.ts
@@ -19,3 +19,9 @@ export interface SpotRequest {
   in_rp_sector_id: string
   log_ids: LogIds
 }
+
+export interface SpotRequestInfo {
+  host: string
+  sector: string
+  salt: string
+}
